First Commit, added everything.
The old repos are now deprecated, besides Unity. I have plans for it.

I have added licenses for the shaders which didn't have them, fixed some to work with new ReShade versions, and even added some others to make this new, including ones that aren't on RetroArch but fits the repo's objective of being a place for Retro-like shaders.